v 10

+++ app/src/main/java/com/aafiyahtech/ventilator/models/Group_3_A.kt
@@ -3,7 +3,7 @@ package com.aafiyahtech.ventilator.models
 import com.google.gson.annotations.SerializedName
 
 data class Group_3_A(
-    @SerializedName("GraphType") val graphType: Float,
-    @SerializedName("AppPollingRate") val appPoleRate: Float,
-    @SerializedName("GraphPollingRate") val graphPolRate: Float
+    @SerializedName("GraphType") val graphType: Int,
+    @SerializedName("AppPollingRate") val appPoleRate: Int,
+    @SerializedName("GraphPollingRate") val graphPolRate: Int
 )
