+++ app/src/main/java/com/aafiyahtech/ventilator/models/Group_6_A.kt
@@ -7,5 +7,5 @@ data class Group_6_A(
     @SerializedName("ActFlow") val actFlow: Float,
     @SerializedName("ActPressure") val actPressure: Float,
     @SerializedName("ActVolume") val actVol: Float,
-    @SerializedName("RefTime") val refTime: Float
+    @SerializedName("RefTime") val refTime: Long
 )
